Rename to reflect that this represents tenant config
