+++ lib/platform/stripes-platform.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const { defaultConfig, emptyConfig, mergeConfig } = require('./platform-config');
+const { defaultConfig, emptyConfig, mergeConfig } = require('./tenant-config');
 const AliasService = require('./alias-service');
 const webpackCommon = require('../webpack-common');
 const logger = require('../cli/logger')();
