+++ lib/platform/tenant-config.js
